Add files via upload
Dishathon Idea: We team fractal propose to create a complete robust platform for the Dish Tv users. The major features we plan to include in the platform: User recommendation using a machine learning algorithm. Censoring of Adult content using image processing and live room tracking to see if a child is at the couch viewing the content Auto pause and record feature that tracks the person's presence in the room and automatically starts recording the content. A very intelligent bot that is enabled with STT and TTS so old people and so many others can request a command to change the channel and so much more just with their voice. Building a super intelligent feedback system based on user experience that generates show by show data to help in user recommendation curation.

So coming to the technologies We aim to build a completely scalable solution that can be deployed in the cloud and will exceed any kind of safety and loading issues. We plan to use the Azure cosmos DB to provide users with close to zero latency in their services.

So the major revolutionary features are:

The automatic child detection mode.So in this mode if the STB/Tv detects the a Child in the room then it will automatically censor the content that is being displayed. So we are talking about realtime censoring and automatic child detection. We plan to do this by sending preprocessed frames and on detection of a child we enable the blur/censor on the screen.